Pulled out the event listener for the Add Task button
